CommentServicev1

Approved-by: Vladimir Radojkovic
Approved-by: Lena

+++ src/main/java/com/example/SocialNetwork/service/FriendsServiceImpl.java
@@ -1,25 +1,31 @@
 package com.example.SocialNetwork.service;
 
+import com.example.SocialNetwork.dto.FriendsDTO;
 import com.example.SocialNetwork.dto.UserDTO;
 import com.example.SocialNetwork.entities.Friends;
 import com.example.SocialNetwork.entities.User;
 import com.example.SocialNetwork.repository.FriendsRepository;
+import com.example.SocialNetwork.repository.UserRepository;
 import org.modelmapper.ModelMapper;
 import jakarta.transaction.Transactional;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FriendsServiceImpl implements FriendsService {
     private ModelMapper mapper;
     private FriendsRepository friendsRepository;
+    private UserRepository userRepository;
 
-    public FriendsServiceImpl(FriendsRepository friendsRepository,  ModelMapper mapper) {
+    public FriendsServiceImpl(FriendsRepository friendsRepository,  ModelMapper mapper, UserRepository userRepository) {
         this.friendsRepository = friendsRepository;
         this.mapper=mapper;
+        this.userRepository=userRepository;
     }
 
     @Override
@@ -29,15 +35,25 @@ public class FriendsServiceImpl implements FriendsService {
     }
 
     @Override
-    public List<UserDTO> getFriendsByUser(Long userId) {
+    public List<UserDTO> getFriendsByUser() {
+        Long userId = getCurrentUser().getId();
         return friendsRepository.getFriendsByUser(userId).stream().map(friend->mapper.map(friend, UserDTO.class)).toList();
     }
 
     @Override
     @Transactional
     public ResponseEntity<String> deleteFriend(Long friendId) {
-        friendsRepository.deleteById(friendId);
-        return new ResponseEntity<>("Friend deleted", HttpStatus.OK);
+        User user = getCurrentUser();
+        List<Friends> friends = user.getFriends();
+        for (Friends friend : friends) {
+            if (friend.getId().equals(friendId)) {
+                friends.remove(friend);
+                userRepository.save(user);
+                friendsRepository.deleteById(friendId);
+                return new ResponseEntity<>("Friend deleted", HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>("Friend not found", HttpStatus.NOT_FOUND);
     }
 
     @Override
@@ -46,5 +62,10 @@ public class FriendsServiceImpl implements FriendsService {
         friendsRepository.deleteFriendByUser(user1Id, user2Id);
         return new ResponseEntity<>("Friend deleted", HttpStatus.OK);
 
+    }
+
+    public User getCurrentUser() {
+        Optional<User> user = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
+        return user.get();
     }
 }
